Refactorización estructura webapp Cambio a arquitectura basada en features/funcionalidades
