+++ webapp/src/features/auth/register/RegisterForm.test.tsx
@@ -1,7 +1,6 @@
 import { describe, expect, it } from 'vitest' 
 import '@testing-library/jest-dom'
 
-
 describe('RegisterForm', () =>
 {
     it('renders without errors', () =>
